+++ flint/prefect/common/ms.py
@@ -7,12 +7,15 @@ from typing import ParamSpec, TypeVar
 
 from prefect import Task, task
 
-from flint.calibrate.aocalibrate import AddModelOptions, add_model
 from flint.imager.wsclean import WSCleanResult
 from flint.logging import logger
+from flint.ms import subtract_model_from_data_column
+from flint.predict.addmodel import AddModelOptions, add_model
 
 P = ParamSpec("P")
 R = TypeVar("R")
+
+task_subtract_model_from_ms = task(subtract_model_from_data_column)
 
 
 # TODO: This can be a dispatcher type function should
